+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [CURRENT] - Development
 ### Features
-- 
+- Add read only/demo mode function if DEMO_MODE is set to yes in hestia.conf
 
 ### Bugfixes
 -